requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-SocketIO==5.3.6
Flask-CORS==4.0.0
python-dotenv==1.0.0
ultralytics==8.0.196
opencv-python==4.8.1.78
Pillow==10.0.1
tqdm==4.66.1
redis==4.6.0
rq==1.15.1
Werkzeug==2.3.7
PyYAML==6.0.1
numpy==1.24.3
matplotlib==3.7.2
python-socketio==5.8.0
eventlet==0.33.3